feat: moved events to the control_mode package, so that they can be included as an input to control modes
