+++ control_mode/include/control_mode/input_collection.hpp
@@ -22,7 +22,7 @@ namespace control_mode
 {
 
 /**
- * @class An interface for whatever provides InputInterfaces to control modes.
+ * An interface for whatever provides InputInterfaces to control modes.
  *
  * You do not own the lifecycle of the InputCollection. Do not hold a reference to the object. Attempting to use an
  * InputCollection outside of ControlMode::configure_inputs() will cause a segfault.
