+++ teleop_core/include/teleop_core/events/EventListenerQueue.hpp
@@ -11,38 +11,16 @@
 // Created by nova on 1/3/70.
 //
 
-#ifndef EVENTLISTENERQUEUE_HPP
-#define EVENTLISTENERQUEUE_HPP
+#ifndef TELEOP_CORE_EVENTLISTENERQUEUE_HPP
+#define TELEOP_CORE_EVENTLISTENERQUEUE_HPP
 
-#include <queue>
-#include <rclcpp/time.hpp>
-
-#include "teleop_core/events/EventListener.hpp"
+#include "control_mode/event/event_listener_queue.hpp"
 
 namespace teleop::internal
 {
-/**
- * Any event listeners to be invoked are added to this queue, and the execution of EventListener::on_event_invoked is
- * deferred until teleop_modular services the queue
- */
-class EventListenerQueue
-{
-public:
-  /**
-   * Clears the queue, after calling on_event_invoked for elements in the queue
-   */
-  void service(const rclcpp::Time & now);
 
-  /**
-   * Adds a listener to the queue to be invoked in the future when serviced.
-   * @param listener A listener to have on_event_invoked called when serviced.
-   */
-  void enqueue(const EventListener::WeakPtr & listener);
-
-private:
-  std::queue<EventListener::WeakPtr> queue_{};
-};
+using EventListenerQueue = control_mode::internal::EventListenerQueue;
 
 }  // namespace teleop::internal
 
-#endif  // EVENTLISTENERQUEUE_HPP
+#endif  // TELEOP_CORE_EVENTLISTENERQUEUE_HPP
